Separated schema helpers and made a few cleaning renames
