+++ src/app/dataset-block/metadata-block/components/event-details/components/common/topics-property/topics-property.component.ts
@@ -9,7 +9,7 @@ import { ChangeDetectionStrategy, Component, Input } from "@angular/core";
 import { MqttTopicSubscription } from "src/app/api/kamu.graphql.interface";
 import { BasePropertyComponent } from "../base-property/base-property.component";
 import { DataSchemaField } from "src/app/interface/dataset-schema.interface";
-import { extractSchemaFieldsFromData } from "src/app/common/helpers/table.helper";
+import { extractSchemaFieldsFromData } from "src/app/common/helpers/data-schema.helpers";
 import { DynamicTableComponent } from "../../../../../../../common/components/dynamic-table/dynamic-table.component";
 import {
     DynamicTableColumnClassEnum,
